Remove header ids before removing newlines
I used the script for a markdown file similar to  this:
````
## header1 {#header1}
a lot of text
## header2 {#header2}
```
Unfortunately, the regex for "Custom header IDs" has matched with the beginning of the header ID at header1 and with the end of header ID at header 2. As a result, the whole section in between has been removed.

Changing the order and remove the header IDs before removing the newlines has solved it. The dot `.` matches only with all chars except the newline. With the change, we can make sure it only matches with custom IDs written in one row.